division: compare the full pointer, not its low 32 bits, against the debug poison
ShiAssert((int)tmp not_eq 0xdddddddd) truncates a 64-bit pointer to its low 32
bits, so a live node whose low half happened to equal MSVC's debug-CRT freed
poison would trip the assert spuriously. Using intptr_t keeps the check exact on
Windows and simply never fires on Linux, which is correct -- glibc does not write
that poison.

Lower severity than the atcbrain sites fixed alongside (that one truncated the
ATC list on a false positive; this only asserts), but the same class, and
spurious assertions are actively harmful here: several real bugs this session were
found by chasing assertions that had been dismissed as noise.

Noted while sweeping: listadt.cpp:211-215 compares the FULL pointer against the
same poison values, so it is harmless dead code on Linux and was left alone.

Verified TE 2 still reaches 3D with no crashes.

+++ src/campaign/camptask/division.cpp
@@ -496,7 +496,12 @@ int DivisionSanityCheck(void)
 
         while (tmp)
         {
-            ShiAssert((int)tmp not_eq 0xdddddddd);
+            // FF_LINUX: intptr_t, not int. (int) truncates a 64-bit pointer to its
+            // low 32 bits, so a live node whose low half happened to equal the MSVC
+            // debug-CRT poison would trip this assert spuriously. Comparing the full
+            // pointer keeps the check exact on Windows and simply never fires here,
+            // which is correct: glibc does not write that poison.
+            ShiAssert((intptr_t)tmp not_eq (intptr_t)0xdddddddd);
             tmp = tmp->next;
         }
     }
